% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram}
\usage{
plot_histogram(
  data,
  bins = 30,
  title = "Histogram",
  xlab = "Value",
  ylab = "Frequency"
)
}
\arguments{
\item{data}{A numeric vector.}

\item{bins}{Number of bins for the histogram. Default is 30.}

\item{title}{Title of the histogram. Default is "Histogram".}

\item{xlab}{Label for the x-axis. Default is "Value".}

\item{ylab}{Label for the y-axis. Default is "Frequency".}
}
\value{
A ggplot2 object representing the histogram.
}
\description{
This function creates a histogram for a given numeric vector.
}
\examples{
plot_histogram(mtcars$mpg, bins = 10, title = "MPG Histogram")
}
