% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter}
\usage{
plot_scatter(x, y, title = "Scatter Plot", xlab = "X", ylab = "Y")
}
\arguments{
\item{x}{A numeric vector for the x-axis.}

\item{y}{A numeric vector for the y-axis.}

\item{title}{Title of the scatter plot. Default is "Scatter Plot".}

\item{xlab}{Label for the x-axis. Default is "X".}

\item{ylab}{Label for the y-axis. Default is "Y".}
}
\value{
A ggplot2 object representing the scatter plot.
}
\description{
This function creates a scatter plot for two numeric vectors.
}
\examples{
plot_scatter(mtcars$wt, mtcars$mpg, title = "Weight vs MPG")
}
